refactor: eliminate duplication by using a common `DatabaseConfig` base class
EPMDJ-11043
