+++ admin-auth/src/main/kotlin/com/epam/drill/admin/auth/config/AuthDatabaseConfig.kt
@@ -15,29 +15,9 @@
  */
 package com.epam.drill.admin.auth.config
 
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-import org.flywaydb.core.Flyway
-import org.jetbrains.exposed.sql.Database
-import org.jetbrains.exposed.sql.transactions.experimental.newSuspendedTransaction
-import javax.sql.DataSource
+import com.epam.drill.admin.common.config.DatabaseConfig
 
-object AuthDatabaseConfig {
-
-    private var database: Database? = null
-    private var dispatcher: CoroutineDispatcher = Dispatchers.IO
-
-    fun init(dataSource: DataSource) {
-        database = Database.connect(dataSource)
-        Flyway.configure()
-            .dataSource(dataSource)
-            .schemas("auth")
-            .baselineOnMigrate(true)
-            .locations("classpath:auth/db/migration")
-            .load()
-            .migrate()
-    }
-
-    suspend fun <T> transaction(block: suspend () -> T): T =
-        newSuspendedTransaction(dispatcher, database) { block() }
-}
+object AuthDatabaseConfig : DatabaseConfig(
+    dbSchema = "auth",
+    schemaMigrationLocation = "classpath:auth/db/migration"
+)
